requirements.txt
tweepy==3.6.0
WTForms==2.1
facebook_sdk==2.0.0
